1) refactored to a single fabric interceptor, 2) added module registrations for HttpInterceptor
