--- Fabric.Authorization.AccessControl/src/app/services/interceptors/fabric-external-idp-search-interceptor.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { FabricExternalIdpSearchInterceptorService } from './fabric-external-idp-search-interceptor.service';
-
-describe('FabricExternalIdpSearchInterceptorService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [FabricExternalIdpSearchInterceptorService]
-    });
-  });
-
-  it('should be created', inject([FabricExternalIdpSearchInterceptorService], (service: FabricExternalIdpSearchInterceptorService) => {
-    expect(service).toBeTruthy();
-  }));
-});
