+++ Fabric.Authorization.AccessControl/src/app/services/interceptors/fabric-http-interceptor.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { FabricHttpInterceptorService } from './fabric-http-interceptor.service';
+
+describe('FabricHttpInterceptorService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [FabricHttpInterceptorService]
+    });
+  });
+
+  it('should be created', inject([FabricHttpInterceptorService], (service: FabricHttpInterceptorService) => {
+    expect(service).toBeTruthy();
+  }));
+});
